requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Azure SDKs
azure-search-documents>=11.6.0
azure-storage-blob>=12.23.0
azure-cosmos>=4.9.0
openai>=1.55.0
# azure-cognitiveservices-speech not needed -- using REST API instead

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Google Drive
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0

# YouTube
yt-dlp>=2024.12.0
youtube-transcript-api>=1.0.0

# Auth
PyJWT[crypto]>=2.8.0

# Utilities
tiktoken>=0.8.0
aiohttp>=3.11.0
pydantic>=2.10.0
requests>=2.31.0
